refactor(utils): extract format utilities from mock-data to lib/format
Production code should not import from mock modules. Moves formatCurrency,
formatPercentage and getTimeAgo into a dedicated lib/format.ts and updates
all consumers. mock-data.ts re-exports for backward compatibility.
