+++ lib/format.ts
@@ -1,0 +1,26 @@
+export function formatCurrency(amount: number): string {
+  return new Intl.NumberFormat('pt-BR', {
+    style: 'currency',
+    currency: 'USD',
+    minimumFractionDigits: 2
+  }).format(amount).replace('US$', '$')
+}
+
+export function formatPercentage(value: number): string {
+  const sign = value >= 0 ? '+' : ''
+  return `${sign}${value}%`
+}
+
+export function getTimeAgo(date: Date): string {
+  const now = new Date()
+  const diffInMinutes = Math.floor((now.getTime() - date.getTime()) / (1000 * 60))
+
+  if (diffInMinutes < 1) return 'Agora mesmo'
+  if (diffInMinutes < 60) return `${diffInMinutes}min atrás`
+
+  const diffInHours = Math.floor(diffInMinutes / 60)
+  if (diffInHours < 24) return `${diffInHours}h atrás`
+
+  const diffInDays = Math.floor(diffInHours / 24)
+  return `${diffInDays}d atrás`
+}
